This commit achieves a number of things. Includes changes are:
- A new Chunks type was created which allows for sorting a group of Chunks with the sort stdlib package
- Processor messages were cleaned up, still not sure what I want them to say
- StartsWithSo processor wasn't working. Fixed it to check chunk.FirstWord instead of chunk.Data
- Most importantly, I decided to do fullText rendering on the server-side. I built a stringer library which will take any number of indices and loop through them adding <span> tags to the fullText. This allows us to pass this output to the frontend and just do styling on the frontend. This was achieved in a new processor called HTMLProcessor.
- Basic color styles added to the frontend for processor types.

As of now, this is all working like I expected. Doing the HTML on the backend is much better than the JS I was trying to think of (athough my approach could be fine in JS too).

What's left to do now is style the application and write some tests. The hard part should be done.